Fix sending URI to an already running app instance
If there is already a running app instance, trying to initialize
logging will fail since the logfile will already be open for writing.

Reorder the program startup such that the check for already running
instance is done before trying to initialize logging.

This also moves the logging initialization after unhandled exception
handler is added so that if similar bugs are introduced in future at
least the user will get an error message.